String format tweak

+++ NSString+Dejal.m
@@ -241,7 +241,7 @@
     }
 
     if (wantSeconds)
-        secondsString = [self stringWithFormat:NSLocalizedString(@"%d %@", @"Seconds remaining"), seconds,
+        secondsString = [self stringWithFormat:NSLocalizedString(@"%ld %@", @"Seconds remaining"), (long)seconds,
             [self dejal_stringWithIntegerValue:seconds zero:nil singluar:secondSingular plural:secondsPlural]];
 
     if (minutes && seconds && wantSeconds && wantMinutes)
@@ -302,7 +302,7 @@
     else if (bytes >= kilobyte)
         temp = [self stringWithFormat:NSLocalizedStringFromTable(@"%.1f KB", @"DejalOpen", @"Kilobytes"), bytes / kilobyte];
     else
-        temp = [self stringWithFormat:NSLocalizedStringFromTable(@"%d bytes", @"DejalOpen", @"Bytes"), bytes];
+        temp = [self stringWithFormat:NSLocalizedStringFromTable(@"%ld bytes", @"DejalOpen", @"Bytes"), (long)bytes];
 
     return temp;
 }
